% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_catch_group.R
\name{combine_catch_group}
\alias{combine_catch_group}
\title{Combine catch group components}
\usage{
combine_catch_group(df)
}
\arguments{
\item{df}{data frame of catch groups}
}
\value{
character string of combined catch groups
}
\description{
Accepts a dataframe which must contain four columns ("species", "life_stage",
"fin_mark", "fate"). These columns are combined into a single string for each
row where the component columns are separated by an underscore "_".
\code{NA} values are forced to 'NA' characters.
}
