% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_db_upload.R
\name{confirm_db_upload}
\alias{confirm_db_upload}
\title{Confirm an analysis_id exists in the database}
\usage{
confirm_db_upload(conn, analysis_id)
}
\arguments{
\item{conn}{An open DBI connection to the creel database.}

\item{analysis_id}{character; session-specific analysis_id created by \code{generate_analysis_lut()}}
}
\value{
Invisibly, TRUE if the analysis_id was found, otherwise FALSE.
}
\description{
Confirm an analysis_id exists in the database
}
