% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{fishery_manager}
\alias{fishery_manager}
\title{Get 'fishery manager' table}
\usage{
fishery_manager(conn, fishery_name = NULL)
}
\arguments{
\item{conn}{A valid database connection from \code{establish_db_con()}}

\item{fishery_name}{Optional character string for pattern matching in analysis_name}
}
\value{
Tibble of fishery information that includes fishery name, dates, and spatial structure (sections and sites).
}
\description{
Simple wrapper to query the vw_fishery_manager table.
}
