% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bss_runtime.R
\name{get_bss_runtime}
\alias{get_bss_runtime}
\title{Summarize BSS model runtime}
\usage{
get_bss_runtime(stan_fit, output_file = NULL)
}
\arguments{
\item{stan_fit}{A 'stanfit' class object}

\item{output_file}{Directory to save output as text file}
}
\description{
Summarize BSS model runtime
}
