% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bss_settings.R
\name{get_bss_settings}
\alias{get_bss_settings}
\title{Summarize priors, sampler settings, and metadata from a stanfit object}
\usage{
get_bss_settings(ecg, ecg_fit, inputs_bss, print = FALSE)
}
\arguments{
\item{ecg}{Estimate catch group (e.g., Steelhead_Adult_UM_Released)}

\item{ecg_fit}{A fitted model (stanfit object)}

\item{inputs_bss}{Inputs of the stan model}

\item{print}{Logical. Return summary to Console}
}
\value{
List with model priors, sampler settings, and metadata
}
\description{
Summarize priors, sampler settings, and metadata from a stanfit object
}
