% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fishery_estimates.R
\name{get_fishery_estimates}
\alias{get_fishery_estimates}
\title{Obtain and format creel estimates for sets of fishery-years}
\usage{
get_fishery_estimates(
  fishery_names,
  years = NULL,
  max_upload_date = TRUE,
  con = NULL
)
}
\arguments{
\item{fishery_names}{Character string of fishery name or vector of character strings.
If \code{years} is provided, \code{fishery_names} is combined with \code{year} to create fishery names.
If \code{years} is not provided, uses this character string or vector of character strings as
the exact fishery names. Try \code{search_fishery_name()} to see available options.}

\item{years}{Integer or vector of integers identifying years of data to pull.
Optional argument, defaults to NULL.}

\item{max_upload_date}{Logical. If TRUE (default), filters to only the most recent upload
date for each fishery_name and catch_group combination. If FALSE, returns all estimates
for each fishery.}

\item{con}{A valid connection to the WDFW PostgreSQL database. If NULL (default),
function will attempt to establish connection. @seealso \code{\link[=establish_db_con]{establish_db_con()}}}
}
\value{
List with two dataframes:
\itemize{
\item \verb{$estimates}: Main estimates data with catch groups, dates, and data grades
\item \verb{$analysis_metadata}: Analysis-level metadata including upload dates and creators
}
}
\description{
Query model estimates from the creel database for specified fisheries and years.
Returns formatted estimate data with catch groups and metadata. Based on approach used in
\code{get_fishery_data()}.
}
\examples{
\dontrun{
# Get latest estimates for a fishery across multiple years
temp <- get_fishery_estimates(
  fishery_names = "Nisqually",
  years = 2021:2023,
  max_upload_date = TRUE
)

# Get all estimates for specific fishery-year combinations
temp <- get_fishery_estimates(
  fishery_names = c("Skagit 2021", "Skagit 2022"),
  max_upload_date = FALSE
)

# Get latest estimates with custom connection
con <- establish_db_con()
temp <- get_fishery_estimates(
  fishery_names = "Cascade",
  years = 2024,
  con = con
)
}
}
\seealso{
Other internal_data: 
\code{\link{establish_db_con}()},
\code{\link{fetch_db_table}()}
}
\concept{internal_data}
