% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_conversion.R
\name{json_conversion}
\alias{json_conversion}
\title{Convert model estimate metadata objects to json format}
\usage{
json_conversion(type, params, analysis_lut)
}
\arguments{
\item{type}{type of metadata to convert to JSON format for storage in the analysis_lut}

\item{params}{user-input parameters defined during the model estimation process}

\item{analysis_lut}{lookup table created during the model estimation process which stores a session-specific analysis_id key and metadata about the analysis.}
}
\value{
Returns the analysis_lut with a new field containing a JSON string of the \code{type}
}
\description{
Convert model estimate metadata objects to json format
}
\examples{
\dontrun{
json_conversion(type = "script")
json_conversion(type = "r_session")
}
}
\seealso{
Other ETL:
\code{\link[=export_estimates]{export_estimates()}},
\code{\link[=finalize_analysis_lut]{finalize_analysis_lut()}},
\code{\link[=map_data_grade]{map_data_grade()}},
\code{\link[=prep_export]{prep_export()}},
\code{\link[=process_estimates_bss]{process_estimates_bss()}},
\code{\link[=process_estimates_pe]{process_estimates_pe()}},
\code{\link[=transform_estimates]{transform_estimates()}},
\code{\link[=update_analysis_comment]{update_analysis_comment()}},
\code{\link[=write_lut]{write_lut()}},
\code{\link[=write_stratum]{write_stratum()}},
\code{\link[=write_total]{write_total()}}
}
\concept{ETL}
