% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_creel_website.R
\name{query_creel_website}
\alias{query_creel_website}
\title{Query the creel website for specific combinations of data}
\usage{
query_creel_website(table_name, query_list, limit = 1e+05)
}
\arguments{
\item{table_name}{Name of table (e.g., catch, interview, effort)}

\item{query_list}{Basically \code{filter()} statements as a list of named items. Item name matches the column name of the table, and then item should either be an atomic to be matched exactly, or a vector for matching multiple options. See details.}

\item{limit}{maximum number of rows to return. Numeric, defaults to 100,000}
}
\value{
tibble of creel data
}
\description{
Queries the website storing the creel data, which uses the Socrata API (e.g., https://dev.socrata.com/docs/queries/where)
}
\details{
The query is focused on filtering, and is designed to be written as an R list, with any number of conditions in it. For example, to return only data in which \code{fishery_name} is "Baker summer sockeye 2022", the query list should include \code{fishery_name = "Baker summer sockeye 2022"}. And to return only data for species "Kokanee" or "Rainbow Trout", the query list should also include \code{species = c("Kokanee", "Rainbow Trout")}.
}
\examples{
dat <- query_creel_website(
  table_name = "catch",
  query_list = list( # each item in list should be named for a column in the table
    fishery_name = "Baker summer sockeye 2022",
    fin_mark = "UM",
    species = c("Kokanee", "Rainbow Trout")
    )
   )
head(dat)
}
\seealso{
Other public_data: 
\code{\link{fetch_dwg}()},
\code{\link{fetch_fishery_names}()},
\code{\link{get_fishery_data}()},
\code{\link{search_fishery_name}()}
}
\concept{public_data}
