% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_helpers.R
\name{raw_catch}
\alias{raw_catch}
\title{Get raw catch data for a fishery}
\usage{
raw_catch(fishery, start_date = NULL, end_date = NULL)
}
\arguments{
\item{fishery}{Fishery name}

\item{start_date}{Optional. First date of window.}

\item{end_date}{Optional. Last date of window.}
}
\value{
Table of raw, unexpanded catch reported from angler interviews.
}
\description{
Get raw catch data for a fishery
}
