% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_progress_report.R
\name{render_progress_report}
\alias{render_progress_report}
\title{Produce a creel progress report and summaries}
\usage{
render_progress_report(fishery_name, bss_used = FALSE)
}
\arguments{
\item{fishery_name}{Identifier which represents the spatiotemporal configuration for a given dataset with associated fishery closures.}

\item{bss_used}{Logical TRUE or FALSE that denotes whether estimates were produced by the Bayesian state-space model. The default option is FALSE.}
}
\description{
Produces a Creel Progress Report for a given fishery from a standardized template script. These reports are primarily used for in-season monitoring of freshwater recreational fisheries and contains high level summaries of sampling data with preliminary estimates of total catch and effort.
}
\examples{
\dontrun{
render_progress_report("Skykomish summer Chinook and gamefish 2025")
render_progress_report("Chehalis winter steelhead 2024-25", bss_used = TRUE)
}
}
