% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_public_helpers.R
\name{search_fishery_name}
\alias{search_fishery_name}
\title{Find fishery names from a partial names}
\usage{
search_fishery_name(fishery_partial)
}
\arguments{
\item{fishery_partial}{Partial fishery name. Can take regular expressions, ignores capitalization.}
}
\value{
Character vector of 'fishery_name' identifiers containing the partial value.
}
\description{
Helps identify options for \code{fetch_dwg()}.
}
\examples{
search_fishery_name("gamefish")
search_fishery_name("Humptulips")
}
\seealso{
Other public_data: 
\code{\link{fetch_dwg}()},
\code{\link{fetch_fishery_names}()},
\code{\link{get_fishery_data}()},
\code{\link{query_creel_website}()}
}
\concept{public_data}
