% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_stratum}
\alias{write_stratum}
\title{Write estimates to stratum estimates table}
\usage{
write_stratum(conn, stratum)
}
\arguments{
\item{conn}{database connection}

\item{stratum}{fine grain estimates}
}
\description{
Write estimates to stratum estimates table
}
