% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_total}
\alias{write_total}
\title{Write Total}
\usage{
write_total(conn, creel_estimates_db)
}
\arguments{
\item{conn}{a valid DBI connection. @seealso \code{\link[=establish_db_con]{establish_db_con()}}}

\item{creel_estimates_db}{a list object containing the standardized model outputs that have been processed by \code{prep_export()} to join certain fields with database lookup tables prior to exportation.}
}
\description{
Write Total
}
\seealso{
Other ETL:
\code{\link[=export_estimates]{export_estimates()}},
\code{\link[=finalize_analysis_lut]{finalize_analysis_lut()}},
\code{\link[=json_conversion]{json_conversion()}},
\code{\link[=map_data_grade]{map_data_grade()}},
\code{\link[=prep_export]{prep_export()}},
\code{\link[=process_estimates_bss]{process_estimates_bss()}},
\code{\link[=process_estimates_pe]{process_estimates_pe()}},
\code{\link[=transform_estimates]{transform_estimates()}},
\code{\link[=write_lut]{write_lut()}},
\code{\link[=write_stratum]{write_stratum()}}
}
\concept{ETL}
