% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_db_upload.R
\name{confirm_db_upload}
\alias{confirm_db_upload}
\title{Confirm database upload}
\usage{
confirm_db_upload(conn, analysis_lut)
}
\arguments{
\item{conn}{Connection to WDFW PostgreSQL database made with DBI-compliant RPostgres package. 'con' created by establish_db_con.R function.}

\item{analysis_lut}{Data frame containing session-specific analysis_id and associated metadata created by generate_analysis_lut.R}
}
\value{
nothing returned.
}
\description{
Confirms upload of model estimates by querying model_analysis_lut for session analysis_id.
}
