% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_db_con.R
\name{establish_db_con}
\alias{establish_db_con}
\title{Establish database connection}
\usage{
establish_db_con(
  conn_type = c("config", "odbc"),
  dsn = "creel_estimates",
  config_path = "config.yml"
)
}
\arguments{
\item{conn_type}{\code{"config"} (default) uses a local \code{config.yml} + \code{RPostgres}; \code{"odbc"} uses a system-configured DSN.}

\item{dsn}{Character string denoting the ODBC domain service name (DSN) to connect to.}

\item{config_path}{File path location of the local 'config.yml' file.}
}
\value{
A \code{DBI} connection to a PostgreSQL database management system. Recommend that this object be named "con".
}
\description{
Establishes a connection to the WDFW PostgreSQL database.
Requires either a  local \code{config.yml} file (default) or a configured ODBC DSN.
The config path uses driver from the \code{RPostgres} package and is preferred for performance, particularly during build database writes.
\strong{Note:} when called in a non-interactive session (scheduled render or automated script), the password prompt will block execution.
}
\seealso{
Other internal_data: 
\code{\link{fetch_db_table}()}
}
\concept{internal_data}
