% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_db_con.R
\name{establish_db_con}
\alias{establish_db_con}
\title{Establish database connection}
\usage{
establish_db_con(
  conn_type = c("odbc", "config"),
  dsn = "creel_estimates",
  config_path = "config.yml"
)
}
\arguments{
\item{conn_type}{Character input denoting either "odbc" or "config" connection type.}

\item{dsn}{Character string denoting the ODBC domain service name (DSN) to connect to.}

\item{config_path}{File path location of the local 'config.yml' file.}
}
\value{
A \code{DBI} connection to a PostgreSQL database management system. Recommend that this object be named "con".
}
\description{
Establishes a connection to the WDFW PostgreSQL database. This process requires either a configured ODBC DSN or a local \code{config.yml} file.
}
\seealso{
Other internal_data: 
\code{\link{fetch_db_table}()},
\code{\link{get_fishery_estimates}()}
}
\concept{internal_data}
