% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_estimates.R
\name{export_estimates}
\alias{export_estimates}
\title{Export creel model estimates}
\usage{
export_estimates(params, analysis_lut, creel_estimates)
}
\arguments{
\item{params}{??}

\item{analysis_lut}{??}

\item{creel_estimates}{??}
}
\value{
??
}
\description{
Primary control function of the export, transform, and load (ETL) process.
Takes standardized model outputs and user-input parameters to export the model estimates appropriately.
}
