% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_estimates.R
\name{export_estimates}
\alias{export_estimates}
\title{Export creel model estimates}
\usage{
export_estimates(params, analysis_lut, creel_estimates, conn = NULL)
}
\arguments{
\item{params}{User-input parameters defined during the model estimation process.}

\item{analysis_lut}{Lookup table created during the model estimation process which stores a session-specific analysis_id key and metadata about the analysis.}

\item{creel_estimates}{List object containing model estimates in a standardized format. Typically passed from 'transform_estimates' function.}

\item{conn}{Database connection object. If NULL (default), a new connection will be established.}
}
\value{
??
}
\description{
Primary control function of the export, transform, and load (ETL) process. Takes standardized model outputs and user-input parameters to export the model estimates appropriately.
}
\seealso{
Other ETL: 
\code{\link{json_conversion}()},
\code{\link{map_data_grade}()},
\code{\link{prep_export}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{transform_estimates}()},
\code{\link{write_lut}()},
\code{\link{write_stratum}()},
\code{\link{write_total}()}
}
\concept{ETL}
