% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dwg.R
\name{fetch_dwg}
\alias{fetch_dwg}
\title{fetch_dwg}
\usage{
fetch_dwg(fishery_name)
}
\arguments{
\item{fishery_name}{Name which represents the spatiotemporal configuration for a given dataset with associated fishery closures.}
}
\value{
dwg, a list object containing a dataframe for each database view downloaded (e.g., effort, interview, catch)
}
\description{
Fetch freshwater recreational fishery creel datasets from Washington's public data warehouse data.wa.gov
}
