% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{fishery_catchgroups}
\alias{fishery_catchgroups}
\title{Get 'fishery_catch_groups' view}
\usage{
fishery_catchgroups(conn, fishery_name = NULL, print = FALSE)
}
\arguments{
\item{conn}{A valid database connection from \code{establish_db_con()}}

\item{fishery_name}{Optional character string for pattern matching in analysis_name}

\item{print}{Logical. If \code{TRUE}, prints all rows to the console. Default \code{FALSE}.}
}
\value{
Tibble of catch groups of interest for a given fishery.
}
\description{
Simple wrapper to query the vw_fishery_manager table.
}
