% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{fishery_lut}
\alias{fishery_lut}
\title{Get fishery lookup table}
\usage{
fishery_lut(conn = NULL)
}
\arguments{
\item{conn}{A valid database connection from \code{establish_db_con()}}
}
\value{
Tibble of fishery names with year, start dates, end dates, and metadata
}
\description{
Simple wrapper to query the fishery_lut table.
}
