% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_conversion.R
\name{json_conversion}
\alias{json_conversion}
\title{Convert model estimate metadata objects to json format}
\usage{
json_conversion(type, params, analysis_lut)
}
\arguments{
\item{type}{Character string; one of "script", "regulations", or "r_session". Evan or Colt: help!}

\item{params}{??}

\item{analysis_lut}{??}
}
\value{
Evan or colt: help!
}
\description{
Convert model estimate metadata objects to json format
}
\examples{
\dontrun{
json_conversion(type = "script")
json_conversion(type = "r_session")
}
}
\seealso{
Other ETL: 
\code{\link{export_estimates}()},
\code{\link{map_data_grade}()},
\code{\link{prep_export}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{transform_estimates}()},
\code{\link{write_lut}()},
\code{\link{write_stratum}()},
\code{\link{write_total}()}
}
\concept{ETL}
