% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data_grade.R
\name{map_data_grade}
\alias{map_data_grade}
\title{Map datagrade ??}
\usage{
map_data_grade(params, transformed_bss_data = NULL, transformed_pe_data = NULL)
}
\arguments{
\item{params}{??}

\item{transformed_bss_data}{??}

\item{transformed_pe_data}{??}
}
\value{
??
}
\description{
Map datagrade ??
}
\seealso{
Other ETL:
\code{\link[=export_estimates]{export_estimates()}},
\code{\link[=finalize_analysis_lut]{finalize_analysis_lut()}},
\code{\link[=json_conversion]{json_conversion()}},
\code{\link[=prep_export]{prep_export()}},
\code{\link[=process_estimates_bss]{process_estimates_bss()}},
\code{\link[=process_estimates_pe]{process_estimates_pe()}},
\code{\link[=transform_estimates]{transform_estimates()}},
\code{\link[=update_analysis_comment]{update_analysis_comment()}}
}
\concept{ETL}
