% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data_grade.R
\name{map_data_grade}
\alias{map_data_grade}
\title{Map datagrade ??}
\usage{
map_data_grade(params, transformed_bss_data = NULL, transformed_pe_data = NULL)
}
\arguments{
\item{params}{??}

\item{transformed_bss_data}{??}

\item{transformed_pe_data}{??}
}
\value{
??
}
\description{
Dev notes: we should avoid \code{assign} here if at all possibles
}
