% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_export.R
\name{prep_export}
\alias{prep_export}
\title{Prep export}
\usage{
prep_export(con, creel_estimates)
}
\arguments{
\item{con}{??}

\item{creel_estimates}{??}
}
\value{
??
}
\description{
Prep export
}
\seealso{
Other ETL: 
\code{\link{export_estimates}()},
\code{\link{json_conversion}()},
\code{\link{map_data_grade}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{transform_estimates}()},
\code{\link{write_lut}()},
\code{\link{write_stratum}()},
\code{\link{write_total}()}
}
\concept{ETL}
