% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_estimates_pe.R
\name{process_estimates_pe}
\alias{process_estimates_pe}
\title{Title}
\usage{
process_estimates_pe(analysis_lut, estimates_pe, params, days)
}
\arguments{
\item{analysis_lut}{??}

\item{estimates_pe}{??}

\item{params}{??}

\item{days}{output of CreelEstimates prep_days().}
}
\value{
??
}
\description{
Title
}
\seealso{
Other ETL:
\code{\link[=export_estimates]{export_estimates()}},
\code{\link[=finalize_analysis_lut]{finalize_analysis_lut()}},
\code{\link[=json_conversion]{json_conversion()}},
\code{\link[=map_data_grade]{map_data_grade()}},
\code{\link[=prep_export]{prep_export()}},
\code{\link[=process_estimates_bss]{process_estimates_bss()}},
\code{\link[=transform_estimates]{transform_estimates()}},
\code{\link[=update_analysis_comment]{update_analysis_comment()}}
}
\concept{ETL}
