% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_estimates.R
\name{transform_estimates}
\alias{transform_estimates}
\title{Transform individual model outputs into a single object}
\usage{
transform_estimates(params, transformed_pe_data, transformed_bss_data)
}
\arguments{
\item{params}{list object containing fishery- and analysis- specific metadata}

\item{transformed_pe_data}{standardized outputs from the PE model, returned by \code{process_pe_estimates()}}

\item{transformed_bss_data}{standardized outputs from the BSS model, returned by \code{process_bss_estimates()}}
}
\value{
list object containing standardized model estimates
}
\description{
Transform individual model outputs into a single object
}
\seealso{
Other ETL:
\code{\link[=export_estimates]{export_estimates()}},
\code{\link[=finalize_analysis_lut]{finalize_analysis_lut()}},
\code{\link[=json_conversion]{json_conversion()}},
\code{\link[=map_data_grade]{map_data_grade()}},
\code{\link[=prep_export]{prep_export()}},
\code{\link[=process_estimates_bss]{process_estimates_bss()}},
\code{\link[=process_estimates_pe]{process_estimates_pe()}},
\code{\link[=update_analysis_comment]{update_analysis_comment()}}
}
\concept{ETL}
