% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_estimates.R
\name{transform_estimates}
\alias{transform_estimates}
\title{Title}
\usage{
transform_estimates(dwg, transformed_pe_data, transformed_bss_data)
}
\arguments{
\item{dwg}{??}

\item{transformed_pe_data}{??}

\item{transformed_bss_data}{??}
}
\value{
??
}
\description{
Title
}
