% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_estimates.R
\name{transform_estimates}
\alias{transform_estimates}
\title{Transform individual model outputs into a single object}
\usage{
transform_estimates(dwg, transformed_pe_data, transformed_bss_data)
}
\arguments{
\item{dwg}{list object containing freshwater creel dataset, returned by \code{fetch_dwg()}}

\item{transformed_pe_data}{standardized outputs from the PE model, returned by \code{process_pe_estimates()}}

\item{transformed_bss_data}{standardized outputs from the BSS model, returned by \code{process_bss_estimates()}}
}
\value{
list object containing standardized model estimates
}
\description{
Transform individual model outputs into a single object
}
\seealso{
Other ETL: 
\code{\link{export_estimates}()},
\code{\link{json_conversion}()},
\code{\link{map_data_grade}()},
\code{\link{prep_export}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{write_lut}()},
\code{\link{write_stratum}()},
\code{\link{write_total}()}
}
\concept{ETL}
