% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_lut}
\alias{write_lut}
\title{Write metadata to analysis lookup table}
\usage{
write_lut(conn, analysis_lut)
}
\arguments{
\item{conn}{database connection}

\item{analysis_lut}{analysis lookup table}
}
\description{
Write metadata to analysis lookup table
}
