% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_lut}
\alias{write_lut}
\title{Write look-up table}
\usage{
write_lut(con, analysis_lut, max_retries = 5)
}
\arguments{
\item{con}{a valid \code{DBI} connection. @seealso \code{\link[=establish_db_con]{establish_db_con()}}}

\item{analysis_lut}{lookup table created during the model estimation process which stores a session-specific analysis_id key and metadata about the analysis.}

\item{max_retries}{maximum number of times to try to write; numeric, defaults to 5.}
}
\description{
Write look-up table
}
\seealso{
Other ETL: 
\code{\link{export_estimates}()},
\code{\link{json_conversion}()},
\code{\link{map_data_grade}()},
\code{\link{prep_export}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{transform_estimates}()},
\code{\link{write_stratum}()},
\code{\link{write_total}()}
}
\concept{ETL}
