% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_analysis_lut.R
\name{write_lut}
\alias{write_lut}
\title{Write look-up table}
\usage{
write_lut(con, analysis_lut, max_retries = 5)
}
\arguments{
\item{con}{??}

\item{analysis_lut}{??}

\item{max_retries}{maximum number of times to try to write; numeric, defaults to 5.}
}
\description{
Write look-up table
}
