% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_stratum}
\alias{write_stratum}
\title{Write stratum}
\usage{
write_stratum(con, creel_estimates_db)
}
\arguments{
\item{con}{a valid DBI connection. @seealso \code{\link[=establish_db_con]{establish_db_con()}}}

\item{creel_estimates_db}{a list object containing the standardized model outputs that have been processed by \code{prep_export()} to join certain fields with database lookup tables prior to exportation.}
}
\description{
Write stratum
}
\seealso{
Other ETL: 
\code{\link{export_estimates}()},
\code{\link{json_conversion}()},
\code{\link{map_data_grade}()},
\code{\link{prep_export}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{transform_estimates}()},
\code{\link{write_lut}()},
\code{\link{write_total}()}
}
\concept{ETL}
