% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_estimates_stratum.R
\name{write_stratum}
\alias{write_stratum}
\title{Write stratum}
\usage{
write_stratum(con, creel_estimates_db)
}
\arguments{
\item{con}{??}

\item{creel_estimates_db}{??}
}
\description{
Write stratum
}
\seealso{
Other ETL: 
\code{\link{export_estimates}()},
\code{\link{json_conversion}()},
\code{\link{map_data_grade}()},
\code{\link{prep_export}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{transform_estimates}()},
\code{\link{write_lut}()},
\code{\link{write_total}()}
}
\concept{ETL}
