% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_estimates_stratum.R
\name{write_stratum}
\alias{write_stratum}
\title{Write stratum}
\usage{
write_stratum(con, creel_estimates_db)
}
\arguments{
\item{con}{??}

\item{creel_estimates_db}{??}
}
\description{
Write stratum
}
