% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_total}
\alias{write_total}
\title{Write estimates to total estimates table}
\usage{
write_total(conn, total)
}
\arguments{
\item{conn}{database connection}

\item{total}{summarized season-total estimates}
}
\description{
Write estimates to total estimates table
}
