% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_estimates_total.R
\name{write_total}
\alias{write_total}
\title{Write Total}
\usage{
write_total(con, creel_estimates_db)
}
\arguments{
\item{con}{??}

\item{creel_estimates_db}{??}
}
\description{
Write Total
}
\seealso{
Other ETL: 
\code{\link{export_estimates}()},
\code{\link{json_conversion}()},
\code{\link{map_data_grade}()},
\code{\link{prep_export}()},
\code{\link{process_estimates_bss}()},
\code{\link{process_estimates_pe}()},
\code{\link{transform_estimates}()},
\code{\link{write_lut}()},
\code{\link{write_stratum}()}
}
\concept{ETL}
